requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
black==19.10b0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
coverage==5.0.3
cryptography==2.8
Django==3.0.2
django-rest-framework==0.1.0
-e git+git@github.com:dcharbon/django-rest-framework-condition.git@d9803dc68acca1c7a4a9ffe831d684f7b33f8c01#egg=django_rest_framework_condition
-e git+git@github.com:Quansight/django-scoped-rbac.git@4bca2940f6000da3dfe5ff1b56adcb38958860dd#egg=django_scoped_rbac
djangorestframework==3.11.0
docutils==0.16
drf-extensions==0.6.0
Faker==4.0.0
github3.py==1.3.0
idna==2.8
importlib-metadata==1.5.0
jeepney==0.4.2
jwcrypto==0.6.0
keyring==21.1.0
lazyasd==0.1.4
more-itertools==8.2.0
ndg-httpsclient==0.5.1
packaging==20.1
pathspec==0.7.0
pkginfo==1.5.0.1
pluggy==0.13.1
prompt-toolkit==3.0.3
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
re-ver==0.4.3
readme-renderer==24.0
regex==2020.1.8
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
-e git+git@github.com:Quansight/safetydance.git@477a013f38acb5beaccd1949aa06b9d5c998e332#egg=safetydance
-e git+git@github.com:Quansight/safetydance_django.git@7295cbb79b1ba428f424f17e998a99815fbabde2#egg=safetydance_django
-e git+git@github.com:Quansight/safetydance_test.git@2b898e5d27e4148f54c897fe0215d6b85cdb7b27#egg=safetydance_test
SecretStorage==3.1.2
setproctitle==1.1.10
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.42.1
twine==3.1.1
-e git+git@github.com:Quansight/type_extensions.git@53b2a65365bf04b14f0d0724f472a8aa899023ae#egg=type_extensions
typed-ast==1.4.1
uritemplate==2.0.0
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
xonsh==0.9.13
zipp==2.1.0
